+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - Use proper Architecture for Debian packages built for musl, see #2811 (@Enselic)
 - Pull in fix for unsafe-libyaml security advisory, see #2812 (@dtolnay)
 - Update git-version dependency to use Syn v2, see #2816 (@dtolnay)
+- Display which theme is the default one, see #2838 (@sblondon)
 - Update git2 dependency to v0.18.2, see #2852 (@eth-p)
 - Improve performance when color output disabled, see #2397 and #2857 (@eth-p)
 - Relax syntax mapping rule restrictions to allow brace expansion #2865 (@cyqsimon)